kernel_drivers: v2: Add in Ingenic JZ4770 support.
Add in Ingenic JZ4770 SoC support. This driver works on the GCW0
platform and gets some 3D rendering working. It should be noted
the version of this driver is 'v2.5.3.2.2.p3' and the latest
Ingenic driver is 'v2.5.3.2.5.b7' for the GCW0. The v2 codebase
will need to be upgraded to this version.

Signed-off-by: Steven J. Hill <sjhill@realitydiluted.com>